Hacks to is_Digit function, and divisibility logic
*Warning. This version does not pass leetcode submit. See below. *

Python's is digit funciton does not recognize negative numbers. The remeby to this was to use  try/except logic to try to cast any input to an int, and except ValueErrors which will alert you that the value is not a digit.

I used the absolute value function to compare inputs for division. More info is in the comments.

*Failing input : ["-78","-33","196","+","-19","-","115","+","-","-99","/","-18","8","*","-86","-","-","16","/","26","-14","-","-","47","-","101","-","163","*","143","-","0","-","171","+","120","*","-60","+","156","/","173","/","-24","11","+","21","/","*","44","*","180","70","-40","-","*","86","132","-84","+","*","-","38","/","/","21","28","/","+","83","/","-31","156","-","+","28","/","95","-","120","+","8","*","90","-","-94","*","-73","/","-62","/","93","*","196","-","-59","+","187","-","143","/","-79","-89","+","-"] *